requirements.txt
# Core dependencies with version pinning for stability
numpy==1.24.4
torch==2.0.1
torchvision==0.15.2
transformers==4.36.2
Pillow==10.0.1
faiss-cpu==1.7.4
streamlit==1.28.1

# Additional utilities
tqdm
